v1.11.8: Success!
Finally, back to where we were...two weeks ago... lol
But seriously, good news. The game is functioning as it was in its weird
OOP-ECS hybrid state, but now in a fully ECS system in which it will be
easier to develop.

Weapon fixes:
Velocity, not torques, are now used
Fixed bug with weapon active state
Fixed bug with weapon knockback
Fixed bug with item not un-equipping properly

Added opacity to quest indicators
Fixed bug with quest indicators not being removed from map